added all tests, fixed naming issues

+++ QtDataSync/sqllocalstore.cpp
@@ -6,6 +6,13 @@
 
 using namespace QtDataSync;
 
+#define TABLE_DIR(tName) \
+	auto tableDir = storageDir; \
+	if(!tableDir.mkpath(tName) || !tableDir.cd(tName)) { \
+		emit requestFailed(id, QStringLiteral("Failed to create table directory %1").arg(tName)); \
+		return; \
+	}
+
 #define EXEC_QUERY(query) do {\
 	if(!query.exec()) { \
 		emit requestFailed(id, query.lastError().text()); \
@@ -39,9 +46,28 @@ void SqlLocalStore::finalize()
 	QSqlDatabase::removeDatabase(DatabaseName);
 }
 
+void SqlLocalStore::count(quint64 id, int metaTypeId)
+{
+	auto tName = tableName(metaTypeId);
+
+	if(testTableExists(tName)) {
+		QSqlQuery countQuery(database);
+		countQuery.prepare(QStringLiteral("SELECT Count(*) FROM %1").arg(tName));
+		EXEC_QUERY(countQuery);
+
+		if(countQuery.first()) {
+			emit requestCompleted(id, countQuery.value(0).toInt());
+			return;
+		}
+	}
+
+	emit requestCompleted(id, 0);
+}
+
 void SqlLocalStore::loadAll(quint64 id, int metaTypeId)
 {
 	auto tName = tableName(metaTypeId);
+	TABLE_DIR(tName)
 
 	if(testTableExists(tName)) {
 		QSqlQuery loadQuery(database);
@@ -50,13 +76,15 @@ void SqlLocalStore::loadAll(quint64 id, int metaTypeId)
 
 		QJsonArray array;
 		while(loadQuery.next()) {
-			QFile file(storageDir.absoluteFilePath(loadQuery.value(0).toString() + QStringLiteral(".dat")));
+			QFile file(tableDir.absoluteFilePath(loadQuery.value(0).toString() + QStringLiteral(".dat")));
 			file.open(QIODevice::ReadOnly);
 			auto doc = QJsonDocument::fromBinaryData(file.readAll());
 			file.close();
 
-			if(doc.isNull()) {
-				emit requestFailed(id, QStringLiteral("Failed to read data from file %1").arg(file.fileName()));
+			if(doc.isNull() || file.error() != QFile::NoError) {
+				emit requestFailed(id, QStringLiteral("Failed to read data from file \"%1\" with error: %2")
+								   .arg(file.fileName())
+								   .arg(file.errorString()));
 				return;
 			} else
 				array.append(doc.object());
@@ -70,6 +98,7 @@ void SqlLocalStore::loadAll(quint64 id, int metaTypeId)
 void SqlLocalStore::load(quint64 id, int metaTypeId, const QString &, const QString &value)
 {
 	auto tName = tableName(metaTypeId);
+	TABLE_DIR(tName)
 
 	if(testTableExists(tName)) {
 		QSqlQuery loadQuery(database);
@@ -78,14 +107,16 @@ void SqlLocalStore::load(quint64 id, int metaTypeId, const QString &, const QStr
 		EXEC_QUERY(loadQuery);
 
 		if(loadQuery.first()) {
-			QFile file(storageDir.absoluteFilePath(loadQuery.value(0).toString() + QStringLiteral(".dat")));
+			QFile file(tableDir.absoluteFilePath(loadQuery.value(0).toString() + QStringLiteral(".dat")));
 			file.open(QIODevice::ReadOnly);
 			auto doc = QJsonDocument::fromBinaryData(file.readAll());
 			file.close();
 
-			if(doc.isNull())
-				emit requestFailed(id, QStringLiteral("Failed to read data from file %1").arg(file.fileName()));
-			else
+			if(doc.isNull() || file.error() != QFile::NoError) {
+				emit requestFailed(id, QStringLiteral("Failed to read data from file \"%1\" with error: %2")
+								   .arg(file.fileName())
+								   .arg(file.errorString()));
+			} else
 				emit requestCompleted(id, doc.object());
 			return;
 		}
@@ -98,6 +129,7 @@ void SqlLocalStore::save(quint64 id, int metaTypeId, const QString &key, const Q
 {
 	auto tName = tableName(metaTypeId);
 	auto tKey = object[key].toVariant().toString();
+	TABLE_DIR(tName)
 
 	//create table
 	QSqlQuery createQuery(database);
@@ -115,17 +147,17 @@ void SqlLocalStore::save(quint64 id, int metaTypeId, const QString &key, const Q
 	EXEC_QUERY(existQuery);
 
 	auto needUpdate = false;
-	QFile *file = nullptr;
+	QScopedPointer<QFile> file;
 	if(existQuery.first()) {
-		file = new QFile(storageDir.absoluteFilePath(existQuery.value(0).toString() + QStringLiteral(".dat")));
+		file.reset(new QFile(tableDir.absoluteFilePath(existQuery.value(0).toString() + QStringLiteral(".dat"))));
 		file->open(QIODevice::WriteOnly);
 	} else {
 		auto fileName = QString::fromLatin1(QUuid::createUuid().toRfc4122().toHex());
-		fileName = storageDir.absoluteFilePath(QStringLiteral("%1XXXXXX.dat")).arg(fileName);
+		fileName = tableDir.absoluteFilePath(QStringLiteral("%1XXXXXX.dat")).arg(fileName);
 		auto tFile = new QTemporaryFile(fileName);
 		tFile->setAutoRemove(false);
 		tFile->open();
-		file = tFile;
+		file.reset(tFile);
 		needUpdate = true;
 	}
 
@@ -133,13 +165,19 @@ void SqlLocalStore::save(quint64 id, int metaTypeId, const QString &key, const Q
 	QJsonDocument doc(object);
 	file->write(doc.toBinaryData());
 	file->close();
+	if(file->error() != QFile::NoError) {
+		emit requestFailed(id, QStringLiteral("Failed to write data to file \"%1\" with error: %2")
+						   .arg(file->fileName())
+						   .arg(file->errorString()));
+		return;
+	}
 
 	//save key in database
 	if(needUpdate) {
 		QSqlQuery insertQuery(database);
 		insertQuery.prepare(QStringLiteral("INSERT INTO %1 (Key, File) VALUES(?, ?)").arg(tName));
 		insertQuery.addBindValue(tKey);
-		insertQuery.addBindValue(storageDir.relativeFilePath(QFileInfo(file->fileName()).completeBaseName()));
+		insertQuery.addBindValue(tableDir.relativeFilePath(QFileInfo(file->fileName()).completeBaseName()));
 		EXEC_QUERY(insertQuery);
 	}
 
@@ -149,6 +187,7 @@ void SqlLocalStore::save(quint64 id, int metaTypeId, const QString &key, const Q
 void SqlLocalStore::remove(quint64 id, int metaTypeId, const QString &, const QString &value)
 {
 	auto tName = tableName(metaTypeId);
+	TABLE_DIR(tName)
 
 	if(testTableExists(tName)) {
 		QSqlQuery loadQuery(database);
@@ -157,7 +196,7 @@ void SqlLocalStore::remove(quint64 id, int metaTypeId, const QString &, const QS
 		EXEC_QUERY(loadQuery);
 
 		if(loadQuery.first()) {
-			auto fileName = loadQuery.value(0).toString() + QStringLiteral(".dat");
+			auto fileName = tableDir.absoluteFilePath(loadQuery.value(0).toString() + QStringLiteral(".dat"));
 			if(!QFile::remove(fileName)) {
 				emit requestFailed(id, QStringLiteral("Failed to delete file %1").arg(fileName));
 				return;
@@ -176,6 +215,13 @@ void SqlLocalStore::remove(quint64 id, int metaTypeId, const QString &, const QS
 void SqlLocalStore::removeAll(quint64 id, int metaTypeId)
 {
 	auto tName = tableName(metaTypeId);
+	TABLE_DIR(tName)
+
+	if(!tableDir.removeRecursively()) {
+		emit requestFailed(id, QStringLiteral("Failed to delete table directory %1").arg(tName));
+		return;
+	}
+
 	QSqlQuery removeQuery(database);
 	removeQuery.prepare(QStringLiteral("DROP TABLE IF EXISTS %1").arg(tName));
 	EXEC_QUERY(removeQuery);
@@ -184,7 +230,7 @@ void SqlLocalStore::removeAll(quint64 id, int metaTypeId)
 
 QString SqlLocalStore::tableName(int metaTypeId) const
 {
-	return QString::fromLatin1(QByteArray(QMetaType::typeName(metaTypeId)).toBase64());
+	return QString::fromLatin1('_' + QByteArray(QMetaType::typeName(metaTypeId)).toHex());
 }
 
 bool SqlLocalStore::testTableExists(const QString &tableName) const
